Move `RepoEntity` and `ModuleEntity` into the `tool` package.
The `model` package is intended to be the data provided by the runtime, independent of any tools consuming that data. These classes are increasingly specific to the tool by tracking cross-references between objects, and surfacing data in a way that's easy to consume by a template system.
