+++ src/main/java/dev/ionfusion/fusion/_private/doc/tool/layout/ModuleWriter.java
@@ -9,8 +9,8 @@ import dev.ionfusion.fusion.ModuleIdentity;
 import dev.ionfusion.fusion._private.StreamWriter;
 import dev.ionfusion.fusion._private.doc.model.BindingDoc;
 import dev.ionfusion.fusion._private.doc.model.ModuleDocs;
-import dev.ionfusion.fusion._private.doc.model.ModuleEntity;
 import dev.ionfusion.fusion._private.doc.tool.MarkdownWriter;
+import dev.ionfusion.fusion._private.doc.tool.ModuleEntity;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
