+++ src/main/java/dev/ionfusion/fusion/cli/Document.java
@@ -8,7 +8,7 @@ import static java.nio.file.Files.isDirectory;
 import com.amazon.ion.Timestamp;
 import dev.ionfusion.fusion.FusionRuntime;
 import dev.ionfusion.fusion.ModuleIdentity;
-import dev.ionfusion.fusion._private.doc.model.RepoEntity;
+import dev.ionfusion.fusion._private.doc.tool.RepoEntity;
 import dev.ionfusion.fusion._private.doc.tool.SiteBuilder;
 import java.io.File;
 import java.io.PrintWriter;
